updated tests for tickets 1, 2 & 3
